+++ src/builder/index.ts
@@ -24,6 +24,8 @@ export type {
   BuildCcdaFunctionalStatusOrganizer,
   BuildCcdaMentalStatus,
   BuildCcdaMentalStatusOrganizer,
+  BuildCcdaAssessmentScale,
+  BuildCcdaAssessmentScaleItem,
   BuildCcdaPlannedItem,
   BuildCcdaPlannedAct,
   BuildCcdaPlannedOrder,